fix test errors

+++ tests/FLang.Tests/Harness/control_flow/break_statement.f
@@ -2,7 +2,7 @@
 //! EXIT: 3
 pub fn main() i32 {
     let count: i32 = 0
-    for (i in 0..10) {
+    for (i in 0..10 as i32) {
         if (i == 3) {
             break
         }
