+++ tests/FLang.Tests/Harness/control_flow/continue_statement.f
@@ -2,7 +2,7 @@
 //! EXIT: 4
 pub fn main() i32 {
     let count: i32 = 0
-    for (i in 0..5) {
+    for (i in 0..5 as i32) {
         if (i == 2) {
             continue
         }
