+++ tests/FLang.Tests/Harness/iterators/iterator_range_basic.f
@@ -5,9 +5,9 @@ import core.range
 
 pub fn main() i32 {
     let sum: i32 = 0
-    let r = 0..4
+    let r = 0..4 as i32
     for (i in r) {
-        sum = sum + i as i32
+        sum = sum + i
     }
     return sum  // 0 + 1 + 2 + 3 = 6
 }
